feat: 模写。TaskContextクラスを用意し、集約
