+++ study-react/src/AddTask.tsx
@@ -1,7 +1,9 @@
 import { useState } from "react";
+import { useTasksDispatch } from "./TaskContext";
 
-export default function AddTask({ onAddTask }) {
+export default function AddTask() {
   const [text, setText] = useState("");
+  const dispatch = useTasksDispatch();
   return (
     <>
       <input
@@ -12,7 +14,11 @@ export default function AddTask({ onAddTask }) {
       <button
         onClick={() => {
           setText("");
-          onAddTask(text);
+          dispatch({
+            type: "added",
+            id: nextId++,
+            text: text,
+          });
         }}
       >
         Add
@@ -20,3 +26,5 @@ export default function AddTask({ onAddTask }) {
     </>
   );
 }
+
+let nextId = 3;
